fix references
