+++ libs/remix-lib/src/helpers/compilerHelper.ts
@@ -1,5 +1,3 @@
-import {Language} from "@remix-project/remix-solidity-ts";
-
 export function compilerInput (contracts) {
   return JSON.stringify({
     language: 'Solidity',
@@ -21,14 +19,4 @@ export function compilerInput (contracts) {
       }
     }
   })
-}
-
-export const Languages = ['Solidity', 'Yul']
-
-export function getValidLanguage (val: string): Language {
-  if (val !== undefined && val !== null && val) {
-    const lang = val.slice(0, 1).toUpperCase() + val.slice(1).toLowerCase()
-    return Languages.indexOf(lang) > -1 ? lang as Language : null
-  }
-  return null
 }
