+++ libs/remix-solidity/src/compiler/compiler-input.ts
@@ -1,6 +1,6 @@
 'use strict'
 
-import { CompilerInput, Source, CompilerInputOptions } from './types'
+import { CompilerInput, Source, CompilerInputOptions, Language } from './types'
 
 export default (sources: Source, opts: CompilerInputOptions): string => {
   const o: CompilerInput = {
@@ -32,3 +32,14 @@ export default (sources: Source, opts: CompilerInputOptions): string => {
   }
   return JSON.stringify(o)
 }
+
+export const Languages = ['Solidity', 'Yul']
+
+export function getValidLanguage (val: string): Language {
+  if (val !== undefined && val !== null && val) {
+    const lang = val.slice(0, 1).toUpperCase() + val.slice(1).toLowerCase()
+    return Languages.indexOf(lang) > -1 ? lang as Language : null
+  }
+  return null
+}
+
