+++ libs/remix-solidity/src/index.ts
@@ -1,6 +1,6 @@
 export { Compiler } from './compiler/compiler'
 export { compile } from './compiler/compiler-helpers'
-export { default as CompilerInput } from './compiler/compiler-input'
+export { default as CompilerInput, getValidLanguage } from './compiler/compiler-input'
 export { CompilerAbstract } from './compiler/compiler-abstract'
 export * from './compiler/types'
 export { promisedMiniXhr, pathToURL, baseURLBin, baseURLWasm, canUseWorker, urlFromVersion } from './compiler/compiler-utils'
