+++ libs/remix-ui/solidity-compiler/src/lib/compiler-container.tsx
@@ -7,7 +7,7 @@ import { canUseWorker, baseURLBin, baseURLWasm, urlFromVersion, pathToURL, promi
 import { compilerReducer, compilerInitialState } from './reducers/compiler'
 import { resetEditorMode, listenToEvents } from './actions/compiler'
 import { OverlayTrigger, Tooltip } from 'react-bootstrap' // eslint-disable-line
-import { helpers } from '@remix-project/remix-lib'
+import { getValidLanguage } from '@remix-project/remix-solidity'
 
 import './css/style.css'
 
@@ -75,7 +75,7 @@ export const CompilerContainer = (props: CompilerContainerProps) => {
           const optimize = params.optimize
           const runs = params.runs as string
           const evmVersion = params.evmVersion
-          const language = helpers.compiler.getValidLanguage(params.language)
+          const language = getValidLanguage(params.language)
 
           return {
             ...prevState,
